amends buttons to react routing and separates buttons to shared reusable components. adds state to render compnents when clicked
